Add files via upload
Updated to include another script that will grab the current patch version from CIG directly and append it to the cert. I did this because previous version was using a precoded patch version that would be added to the cert. If CIG updates the game it would most likely break this CertFix. This problem has been solved now.